+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -14,7 +14,6 @@ from knack.log import get_logger
 
 from azure.cli.core.commands.validators import validate_tag
 from azure.cli.core.util import CLIError
-from azure.cli.core.azclierror import InvalidArgumentValueError
 import azure.cli.core.keys as keys
 
 from .vendored_sdks.azure_mgmt_preview_aks.v2020_11_01.models import ManagedClusterPropertiesAutoScalerProfile
@@ -247,7 +246,7 @@ def _validate_subnet_id(subnet_id, name):
         return
     from msrestazure.tools import is_valid_resource_id
     if not is_valid_resource_id(subnet_id):
-        raise InvalidArgumentValueError(name + " is not a valid Azure resource ID.")
+        raise CLIError(name + " is not a valid Azure resource ID.")
 
 
 def validate_load_balancer_outbound_ports(namespace):