+++ src/connectedk8s/HISTORY.rst
@@ -5,7 +5,8 @@ Release History
 
 0.2.8
 ++++++
-* Added checks for proxy and disable proxy
+* Added checks for proxy and added disable-proxy
+* Updated config dataplane endpoint to support other clouds
 
 0.2.7
 ++++++